Revert "Confused Deputy Protection for Resource Policies" (#1565)
