+++ deprecated.txt
@@ -37,4 +37,3 @@ Snowflake.PublicRoleGrant
 GCP.K8S.Pot.Create.Or.Modify.Host.Path.Volume.Mount
 Snowflake Attempted Login With Disabled User
 Snowflake.Stream.AttemptedLoginByDisabledUser
-AWS.S3.Bucket.PolicyConfusedDeputyProtection
